+++ main_client/client.go
@@ -6,7 +6,8 @@ import (
 	"time"
 
 	auth "github.com/backend-ids/authentication"
-	"github.com/backend-ids/src/client"
+	auth_client "github.com/backend-ids/src/client/auth_client"
+	ids_client "github.com/backend-ids/src/client/ids_client"
 	pb "github.com/backend-ids/src/proto"
 	_ "github.com/lib/pq"
 	"google.golang.org/grpc"
@@ -35,7 +36,8 @@ func main() {
 		log.Fatalf("Did not connect: %v", err)
 	}
 	defer conn.Close()
-	client2 := client.NewAuthClient(conn, "user-1", "123456")
+	//client2 --->auth_client
+	client2 := auth_client.NewAuthClient(conn, "user-1", "123456")
 	token, err := client2.Login()
 	if err != nil {
 		log.Panic(err)
@@ -52,14 +54,15 @@ func main() {
 	if err != nil {
 		log.Println("cannot create auth interceptor:", err)
 	}
+	//client1 ---> ids_client
 	client1 := pb.NewIdsCRUDClient(clientConnection)
-	// client.CreateQuestion(client1, "subject-1", "question description", 1)
-	client.EditQuestion(client1, 1, "subject-1", "modified description", 1, 2, 1)
-	// client.CreateSolution(client1, "slution 3", 3, 3)
-	// client.EditSolution(client1, 2, "another approach", 2, 3)
-	// client.CreateComment(client1, "ok will do", 1, 3)
-	// client.GetQuestions(client1, 1)
-	// client.GetQuestionById(client1, 1, 2)
-	// client.CreateUser(client1, "user-5", "email-5@email.com", "asdqw#$f123@G", "mentor", "")
-	// client.UpdateUserDetails(client1, 5, "user-5", "email-5@email.com", "123456", "mentor", "")
+	// ids_client.CreateQuestion(client1, "subject-1", "question description", 1)
+	ids_client.EditQuestion(client1, 1, "subject-1", "modified description", 1, 2, 1)
+	// ids_client.CreateSolution(client1, "slution 3", 3, 3)
+	// ids_client.EditSolution(client1, 2, "another approach", 2, 3)
+	// ids_client.CreateComment(client1, "ok will do", 1, 3)
+	// ids_client.GetQuestions(client1, 1)
+	// ids_client.GetQuestionById(client1, 1, 2)
+	// ids_client.CreateUser(client1, "user-5", "email-5@email.com", "asdqw#$f123@G", "mentor", "")
+	// ids_client.UpdateUserDetails(client1, 5, "user-5", "email-5@email.com", "123456", "mentor", "")
 }
